+++ src/Resources/contao/config/config.php
@@ -44,6 +44,17 @@ $GLOBALS['TL_WRAPPERS']['start'][]     = 'bs_gridStart';
 $GLOBALS['TL_WRAPPERS']['separator'][] = 'bs_gridSeparator';
 $GLOBALS['TL_WRAPPERS']['stop'][]      = 'bs_gridStop';
 
+// Hooks
+$GLOBALS['TL_HOOKS']['exportTheme'][] = [
+    'contao_bootstrap.grid.listeners.theme_export',
+    'onExportTheme'
+];
+
+$GLOBALS['TL_HOOKS']['extractThemeFiles'][] = [
+    'contao_bootstrap.grid.listeners.theme_import',
+    'onExtractThemeFiles'
+];
+
 // Easy Themes
 $GLOBALS['TL_EASY_THEMES_MODULES']['bs_grid'] = [
     'href_fragment' => 'table=tl_bs_grid',